+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This is the rolling changelog for TShock for Terraria. Use past tense when addin
 ## Upcoming Changes
 * New permission `tshock.tp.pylon` to enable teleporting via Teleportation Pylons (@QuiCM)
 * New permission `tshock.journey.research` to enable sharing research via item sacrifice (@QuiCM)
+* Add Emoji event to GetDataHandler. This packet is received when a player tries to display an emote.
+	* Adding EmojiHandler to handle an exploit. Adding `tshock.sendemoji` permission and checks. Added this permission to guest group by default.
 
 ## TShock 4.4.0 (Pre-release 10)
 * Fix all rope coils. (@Olink)